Extract downloadJSON to seperate component
