+++ src/modules/auth/components/SetPasswordSuccess/SetPasswordSuccess.css
@@ -32,22 +32,6 @@
     color: var(--color-blue-gray);
   }
 
-  & > .downloadLisk {
-    display: flex;
-    width: 100%;
-    padding: 16px 0;
-    justify-content: center;
-
-    & > p {
-      @mixin contentLarge bold;
-
-      color: var(--color-maastricht-blue);
-      margin: 20px;
-      text-align: left;
-      word-spacing: 8px;
-    }
-  }
-
   & > .continueButton {
     @mixin contentNormal;
