+++ src/modules/auth/components/SetPasswordSuccess/SetPasswordSuccess.js
@@ -1,18 +1,13 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { PrimaryButton, TertiaryButton } from 'src/theme/buttons';
-import { downloadJSON } from '@transaction/utils';
+import { PrimaryButton } from 'src/theme/buttons';
 import Box from 'src/theme/box';
+import DownloadJSON from 'src/modules/common/components/DownloadJSON/DownloadJSON';
 import BoxContent from 'src/theme/box/content';
-import Icon from 'src/theme/Icon';
 import styles from './SetPasswordSuccess.css';
 
 function SetPasswordSuccess({ onClose, encryptedPhrase, buttonText }) {
   const { t } = useTranslation();
-
-  const onDownload = () => {
-    downloadJSON(encryptedPhrase, 'encrypted_secret_recovery_phrase');
-  };
 
   const onContinue = () => onClose();
 
@@ -25,20 +20,7 @@ function SetPasswordSuccess({ onClose, encryptedPhrase, buttonText }) {
             'You can now download your encrypted secret recovery phrase and use it to add your account on other devices.',
           )}
         </p>
-        <div className={styles.downloadLisk}>
-          <Icon name="fileOutline" />
-          <p className="option-value">encrypted_secret_recovery_phrase.json</p>
-          <TertiaryButton
-            className={styles.downloadBtn}
-            size="xs"
-            onClick={onDownload}
-          >
-            <span className={styles.buttonContent}>
-              {t('Download')}
-              <Icon name="downloadBlue" />
-            </span>
-          </TertiaryButton>
-        </div>
+        <DownloadJSON fileName="encrypted_secret_recovery_phrase" encryptedPhrase={encryptedPhrase} />
         <PrimaryButton className={styles.continueButton} onClick={onContinue}>
           { buttonText || t('Continue to Dashboard')}
         </PrimaryButton>
